+++ test/methods/macro_test.js
@@ -2,7 +2,7 @@
 
 module.exports = (it, expect, collect) => {
   it('should be able to register a custom macro/method', () => {
-    collect().macro('uppercase', function () {
+    collect().macro('uppercase', function uppercase() {
       return this.map(item => item.toUpperCase());
     });
 
@@ -10,9 +10,7 @@ module.exports = (it, expect, collect) => {
     expect(collection.uppercase().all()).to.eql(['A', 'B', 'C']);
     expect(collection.all()).to.eql(['a', 'b', 'c']);
 
-    collect().macro('prefix', function (prefix) {
-      return this.map(item => prefix + item);
-    });
+    collect().macro('prefix', prefix => this.map(item => prefix + item));
 
     expect(collect(['a', 'b', 'c']).prefix('xoxo').all()).to.eql(['xoxoa', 'xoxob', 'xoxoc']);
   });